// backend/price_pooler/src/types.ts
export interface BinanceTrade {
  t: string;
  T: number;
  s: string;
  p: string;
  q: string;
  m: boolean;
}

export interface TradeData {
  tradeId: string;
  tradeTime: Date;
  symbol: string;
  priceValue: bigint;
  priceDecimals: number;
  quantityValue: bigint;
  quantityDecimals: number;
  side: string;
}

export interface redisTradeData {
  tradeId : string;
  tradeTime : Date;
  symbol : string;
  price : bigint;
  priceDecimals : number;
  quantity : bigint;
  quantityDecimals : number;
  side : string;
  bid : bigint;
  bidDecimals : number;
  ask : bigint;
  askDecimals : number;
}